pyproject.toml
[tool.poetry]
name = "investmentadvisor"
version = "0.1.0"
description = "A POC for analyzing Stock's annual report PDF using LLM"
authors = ["Apoorv Jain <4poorv@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.172"
streamlit = "^0.89.0"
python-dotenv = "^0.19.0"
chardet = "^4.0.0"
altair = "^4.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
